+++ src/main/kotlin/di/UseCaseModule.kt
@@ -2,9 +2,11 @@ package org.example.di
 
 import logic.usecase.*
 import org.example.logic.usecase.filter.SeaFoodMealUseCase
+import org.example.logic.usecase.game.GuessMealGameUseCase
 import org.example.logic.usecase.game.IngredientGameUseCase
 import org.example.logic.usecase.suggest.EggFreeSweetsUseCase
 import org.example.logic.usecase.suggest.IraqiMealsIdentifierUseCase
+import org.example.logic.usecase.suggest.MealsForLargeGroupUseCase
 import org.example.logic.usecase.suggest.SuggestHighCalorieMealUseCase
 import org.koin.dsl.module
